pyproject.toml
[project]
name = "ragssistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.7.0",
    "chromadb>=1.0.12",
    "hf-xet>=1.1.3",
    "huggingface-hub>=0.31.2",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "langchain-mistralai>=0.2.10",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.5",
    "pydantic>=2.11.4",
    "pypdf>=5.5.0",
    "sentence-transformers>=4.1.0",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
